Queues now wait for the file where they are instantiated to finish parsing before attempting to load any outstanding dependencies.
This gives a module in the same file but later than a queue
which depends upon it the chance to instantiate itself - which
will prevent the queue from trying to load it from a remote source.

The idea is that this allows multiple modules to be included in the
same compressed file for client side applications.

Kicked up version.